glyphsieve/pyproject.toml
[project]
name = "glyphsieve"
version = "0.1.0"
description = "ETL"
readme = "README.md"
authors = [
    { name = "Lucien Brule", email = "lucien@lucienbrule.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "orjson>=3.11.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "rapidfuzz>=3.13.0",
    "sentence-transformers>=5.0.0",
    "scikit-learn>=1.7.1",
    "joblib>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml",
]

[project.scripts]
glyphsieve = "glyphsieve.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "literal"

[tool.setuptools.package-data]
"glyphsieve.resources" = ["*.yaml", "*.json"]
